Create titanic dataset
 you will take the role of a researcher working on a database with the passengers of the Titanic ship which sunk. You are trying to understand what patterns there are in the passengers who survived, and you want to answer the question: what predicted that a passenger would survive?

Please come up with 3 hypothesis of what might predict that a passenger will survive, and show your work. It is ok if you find some things that do not predict at all, or that negatively predict that a passenger will survive. For each hypothesis, make explicit what your hypothesis or question is, make a plot that tries to answer that hypothesis, and explicitely describe your conclusion (i.e. was your hypothesis correct? what is the answer to your question?).